requirements.txt
llama-index>=0.11
llama-index-core>=0.11
llama-index-embeddings-ollama>=0.2
llama-index-llms-ollama>=0.2
pypdf>=4.3.1
python-docx>=1.1.2
unstructured>=0.15.0
unstructured[pdf]
tiktoken>=0.7.0
chromadb>=0.5.5               # optional (not used in v1 but handy if you switch)
#streamlit>=1.37.0
python-dotenv>=1.0.1
langchain==0.3.18
langchain_experimental==0.3.4
openai==1.59.5
pandas==2.2.2
streamlit_chat==0.1.1
streamlit_feedback==0.1.4
tabulate
streamlit==1.28.2
